2023.4
Updates

* SDK Samples now prompt for user input rather than using hardcoded mod IDs
* Consumers can now define `MODIO_NO_DEPRECATED` to completely remove all deprecated functionality in the SDK
* Support added for new PotentiallyHarmful value in the VirusStatus enum. This flag is set for files that did not return a positive virus scan, but contain executables.
* Improved support for HTTP redirection on Linux, which fixes issues relating to some files not correctly downloading.
* Added `FilesizeUncompressed` field to ModFiles. `FilesizeUncompressed` is used to ensure that there is enough free disk space on the device before before extraction begins. A mod installation will return an error if there is insufficient space - a `Modio::FilesystemError::InsufficientSpace` error will be returned in ModManagement callbacks.
* Added new `Modio::GetGameInfoAsync` call to get the `GameInfo` object for a given Game ID. This can be used to retrieve various game-related properties defined on the mod.io website, meaning tooling and game features can be more easily made consistent with data defined in your game profile.
* Improved large file uploads to ensure already uploaded parts are not re-attempted. This helps improve resumption of large file uploads via multipart uploads.

Bugfixes

* `ghc::filesystem` updated to fix a long path issue on Windows
* ModInfo cache will be invalidated when a user submits edits to a mod, so that subsequent calls to `Modio::GetModInfoAsync` return updated results
* File and folder deletion operations now have a cancellation signal on Windows, fixing a bug in `Modio::ShutdownAsync` while a deletion was in progress

Deprecations

* `Modio::Shutdown` (the blocking variant) had a number of issues and was not suitable for use on consoles due to platform requirements and has now been deprecated in favor of `Modio::ShutdownAsync`

Removed Functionality

* All of the ModProgress functionality (deprecated in 2023.1) has now been removed from the SDK. If you were using this deprecated functionality, you need to update to the new Mod Progress methods.

+++ modio/modio/detail/ops/compression/zip/FinalizeArchiveOp.h
@@ -86,9 +86,9 @@ namespace Modio
 								.FollowedBy<std::uint16_t>(0) // Last modified date
 								.FollowedBy<std::uint32_t>(CurrentArchiveEntry->CRCValue)
 								.FollowedBy<std::uint32_t>(ExtraFieldMax ? Constants::ZipTag::MAX32
-																		 : CurrentArchiveEntry->CompressedSize)
+																		 : static_cast<std::uint32_t>(CurrentArchiveEntry->CompressedSize))
 								.FollowedBy<std::uint32_t>(ExtraFieldMax ? Constants::ZipTag::MAX32
-																		 : CurrentArchiveEntry->UncompressedSize)
+																		 : static_cast<std::uint32_t>(CurrentArchiveEntry->UncompressedSize))
 								.FollowedBy<std::uint16_t>(
 									(std::uint16_t) CurrentArchiveEntry->FilePath.generic_u8string().size())
 								.FollowedBy<std::uint16_t>(
@@ -101,7 +101,7 @@ namespace Modio
 								.FollowedBy<std::uint32_t>(
 									// Offset of local file header from start of disk (only one disk, so offset from
 									// start of file)
-									ExtraFieldMax ? Constants::ZipTag::MAX32 : LocalFileOffset);
+									ExtraFieldMax ? Constants::ZipTag::MAX32 : static_cast<std::uint32_t>(LocalFileOffset));
 						}
 
 						{
@@ -176,15 +176,15 @@ namespace Modio
 						.FollowedBy<std::uint16_t>(
 							ArchiveFile->GetNumberOfEntries()) // Number of entries in the central directory
 						.FollowedBy<std::uint32_t>(ArchiveFile->bIsZip64 == true
-													   ? SizeCentralDir
-													   : OutputFile->Tell() -
-															 StartOfCentralDirectory) // Size of central directory
+													   ? static_cast<std::uint32_t>(SizeCentralDir)
+													   : static_cast<std::uint32_t>(OutputFile->Tell() -
+															 StartOfCentralDirectory)) // Size of central directory
 						// For some reason, the macOS decompressor does not like tp have #Entries and Size of CD to
 						// "0xFF" when not necessary
 						.FollowedBy<std::uint32_t>(
 							ArchiveFile->bIsZip64 == true
 								? Constants::ZipTag::MAX32
-								: StartOfCentralDirectory) // Position of central directory in file
+								: static_cast<std::uint32_t>(StartOfCentralDirectory)) // Position of central directory in file
 						.FollowedBy<std::uint16_t>(0); // Length of comment
 
 					yield OutputFile->WriteAsync(std::move(*RecordBuffer), std::move(Self));
